feat(workspace): delete sera-events, migrate call sites to sera-telemetry (P0-2)
Absorbs sera-events modules (audit hash chain, Centrifugo client, channel
namespaces) into sera-telemetry and removes the sera-events crate.

- sera-telemetry: adds hash_chain, centrifugo, channels modules with
  top-level re-exports. Pulls in reqwest + jsonwebtoken workspace deps.
- sera-gateway: swaps sera_events:: imports for sera_telemetry:: across
  state, services/{audit,intercom,notification_service}, and one doc
  comment in routes/intercom.
- sera-meta: drops unused sera-events Cargo dep (no src usage).
- Workspace: removes crates/sera-events from members and the workspace
  dep map; deletes crates/sera-events/.
- CLAUDE.md: strikes sera-events from the crate map + dep graph and
  extends the sera-telemetry row.

Verification:
- cargo check --workspace: green
- cargo test -p sera-gateway: 230 passed, 3 ignored
- cargo test -p sera-telemetry: 119 passed
- cargo clippy --workspace -- -D warnings: clean
