--- rust/crates/sera-events/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "sera-events"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-sera-types.workspace = true
-sera-errors.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-tokio.workspace = true
-tracing.workspace = true
-thiserror.workspace = true
-uuid.workspace = true
-time.workspace = true
-reqwest.workspace = true
-jsonwebtoken.workspace = true
-sha2 = "0.10"
