+++ rust/crates/sera-gateway/Cargo.toml
@@ -17,7 +17,7 @@ rusqlite.workspace = true
 sera-types.workspace = true
 sera-db.workspace = true
 sera-auth.workspace = true
-sera-events.workspace = true
+sera-telemetry.workspace = true
 sera-tools.workspace = true
 sera-config.workspace = true
 sera-errors.workspace = true
