+++ rust/crates/sera-meta/Cargo.toml
@@ -8,7 +8,6 @@ license.workspace = true
 [dependencies]
 sera-types = { workspace = true }
 sera-auth = { workspace = true }
-sera-events = { workspace = true }
 sera-errors = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
